requirements.txt
python-telegram-bot==21.9
yt-dlp
gallery-dl
aiofiles==24.1.0
python-dotenv==1.0.1